backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
sqlalchemy>=2.0.25
alembic>=1.13.1
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
stripe>=7.9.0
ezdxf>=1.4.0
reportlab>=4.0.8
Pillow>=11.0.0
numpy>=1.26.3
aiofiles>=23.2.1
httpx>=0.26.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
shapely>=2.0.0
pypdf>=5.0.0
pdfminer.six>=20240000
jinja2>=3.1.3
aiosmtplib>=3.0.1
slowapi>=0.1.9
structlog>=24.1.0
